When long text is fed to oled.log(), wrap it.
It ignores space breaks for now, sacrificing legibility for compactness.

Things that could be done:
let it take an optional flag that is wrap vs truncate, for people invoking via code.
let it take an optional flag that is wrap-at-width vs wrap-on-space

They could be held in the config to set the defaults, but then overridden on each invocation?

Also not sure whether this will freak out the microprocessor/memory usage since it's implemented (lazily)
via recursion. Maybe a loop is better?