PR branch for little plugin loader.
This is very useful to add persistant code to your badge, while retaining the functionlity of the boot application. You can extend or add functionality,
test random code snippits and experiments in a persistant but easily revertable way.

I admit this may be more useful for a 'always-on' wearable like a badge, and not all Aiko implementations... so if it's not acceptable globally in main.py, let me know and I'll refactor the patch :)