Fix WS2812B dim support between code and MQTT
dim global variable in led.py was configured to have a separate
value in the aiko net MQTT thread, and the main thread.
This means that example code calling lib/led.py to set pixels,
got a different dimming value than the same pixels set by the same
led.py when they were set from the MQTT thread.

Workaround is to store the dim value in configuration/main.py which
is imported in memory before the threads happen, and that namespace
is effectively shared between the different threads, so it can be
used to share data.
Using configuration/led.py would not work as it is not imported
in memeory before aiko forks off the net/MQTT thread.